GPII-2235 : Adding two guards to avoid the duplication of the GPII-App.
First, detect a second instance using app.makeSingleInstance from Electron
to avoid the execution of another instance of the electron app. Second, it
uses the new gpii.SingleInstance capability to detect if a current GPII
FlowManager is running in the system. Quit the application if any of that
events happens.