GPII-2206 : Obtaining the windows PS1 scripts from github directly.
We find a chicken-egg problem, that we need to boostrap fist launching
the correct set of provisioning scripts before call npm install.
Problem is that we haven't gpii-windows before npm install so we need to
obtain directly the scripts from the repository. This method needs to be
optimized and defined as a global method in main library.

Disconnecting the current call to gpii-windows build script because we
are taking care of that directly at the end of the gpii-app build script
with npm.