GPII-3191: Copy the resulting installer into the project folder
The installer will be copied from the "installer" folder into the
original project folder. For instance, on CI, the installer is being
created under c:/installer and the git project is in c:/vagrant. With
this change the installer will be copied from the c:/installer/output
folder into c:/vagrant/installer.

This allows us to publish the installer as an output build artifact.