NOJIRA: Removing the "onRepeaterCreated" and "onHandlerCreated" events
The idea behind those 2 events was to have a reliable point in time when
the elementRepeater was created. However, it turns out that this event
is fired before the "onCreate" event of the repeater. That is why the
"onCreate" event can be used for this purpose and there is no need to
have the "onRepeaterCreated" and "onHandlerCreated" event (the dynamic
handler subcomponents are created via an array and not via an events and
thus the wrapping component will be ready when all of its child
components have been successfully initialized).